+++ cmd/goatcounter/serve.go
@@ -50,9 +50,9 @@ with -dev.
 
 Flags:
 
-  -db          Database connection: "sqlite://<file>" or "postgres://<connect>"
+  -db          Database connection: "sqlite+<file>" or "postgres+<connect>"
                See "goatcounter help db" for detailed documentation. Default:
-               sqlite://db/goatcounter.sqlite3?_busy_timeout=200&_journal_mode=wal&cache=shared
+               sqlite+/db/goatcounter.sqlite3?_busy_timeout=200&_journal_mode=wal&cache=shared
 
   -listen      Address to listen on. Default: "*:443", or "localhost:8081" with
                -dev. See "goatcounter help listen" for detailed documentation.
@@ -257,7 +257,7 @@ func doServe(ctx context.Context, db zdb.DB,
 
 func flagsServe(f zli.Flags, v *zvalidate.Validator) (string, bool, bool, string, string, string, error) {
 	var (
-		dbConnect   = f.String("sqlite://db/goatcounter.sqlite3", "db").Pointer()
+		dbConnect   = f.String("sqlite+db/goatcounter.sqlite3", "db").Pointer()
 		debug       = f.String("", "debug").Pointer()
 		dev         = f.Bool(false, "dev").Pointer()
 		automigrate = f.Bool(false, "automigrate").Pointer()